+++ src/io/win/fs.c
@@ -29,7 +29,7 @@ btc_timespec_set_filetime(btc_timespec_t *ts, const FILETIME *ft) {
   ul.LowPart = ft->dwLowDateTime;
   ul.HighPart = ft->dwHighDateTime;
 
-  ns = (uint64_t)(ul.QuadPart - epoch) * 100;
+  ns = (ul.QuadPart - epoch) * 100;
 
   ts->tv_sec = ns / UINT64_C(1000000000);
   ts->tv_nsec = ns % UINT64_C(1000000000);
